// packages/cc/src/lib/_Types.d.ts
export declare enum NodeNamingAndLocationCommand {
    NameSet = 1,
    NameGet = 2,
    NameReport = 3,
    LocationSet = 4,
    LocationGet = 5,
    LocationReport = 6
}